Struct elma_builder::channel::Channel

pub struct Channel {
    pub _name: String,
    pub _capacity: usize,
    pub _queue: VecDeque<f64>,
}

Arguments

_name - A string holding the name of the channel instantiation.

_capacity - This contains the total capacity of the channel.

_queue - A queue where all data is placed in and read out from the channel.

Remarks

All ELMA channels are meant to have the same fields. For this reason, all Channel fields are described in the Channel struct.

Fields
_name: String_capacity: usize_queue: VecDeque<f64>
Methods

impl Channel

Remarks

The implementation of Channel shares fields from the pub struct Channel. This means that when a channel object is created the object contains fields from the struct and one may operate on the object using the impl functions. The given functions allow for full read/write access to the _capacity and _queue, but the _name is read only.

pub fn size(&self) -> usize

pub fn empty(&self) -> bool

pub fn nonempty(&self) -> bool

pub fn name(&self) -> &String

pub fn capacity(&self) -> &usize

pub fn send(&mut self, value: f64)

pub fn flush(&mut self)

pub fn change_capacity(&mut self, new_cap: usize)

pub fn latest_db(&self) -> f64

pub fn latest_vec(&self, n: usize) -> Vec<f64>

pub fn earliest(&self) -> f64