Struct elma_builder::reciever_proc::Reciever

pub struct Reciever {
    pub _sum: f64,
    pub _period: Duration,
    pub _previous_update: Duration,
    pub _last_update: Duration,
    pub _start_time: SystemTime,
    pub _name: String,
    pub _num_updates: i64,
    pub _status: StatusEnum,
}

Arguments

_sum - Grabs the sum of the data in a channel added to the manager.

_period - The period at which the manager should call the update function of the process.

_previous_update - This contains the the time that the previous update occured as is used to measure when the next update should occur in the manager.

_last_update - Contains the time when the latest update happened.

_start_time - Contains the time when the start function was called for the process.

_name - This contains the name of the process.

_num_updates - A counter for the number of times the process has been updated since instantiation.

_status - Provides an unsigned byte showing the current state of the process. This enum is defined in the process file next to the trait.

Remarks

The reciever looks for data in a data channel given to the manager. If there is data in the channel, it looks at all of it and finds the sum of all of it.

Fields
_sum: f64_period: Duration_previous_update: Duration_last_update: Duration_start_time: SystemTime_name: String_num_updates: i64_status: StatusEnum